+++ hooks/emitter/oracle.go
@@ -257,78 +257,3 @@ func (h *Hook) handleEventDeactivate(ctx sdk.Context, evMap common.EvMap) {
 	h.emitUpdateValidatorStatus(ctx, addr)
 	h.emitHistoricalValidatorStatus(ctx, addr)
 }
-
-func decodeMsgRequestData(msg *types.MsgRequestData, msgJson common.JsDict) {
-	msgJson["oracle_script_id"] = msg.GetOracleScriptID()
-	msgJson["calldata"] = msg.GetCalldata()
-	msgJson["ask_count"] = msg.GetAskCount()
-	msgJson["min_count"] = msg.GetMinCount()
-	msgJson["client_id"] = msg.GetClientID()
-	msgJson["fee_limit"] = msg.GetFeeLimit()
-	msgJson["prepare_gas"] = msg.GetPrepareGas()
-	msgJson["execute_gas"] = msg.GetExecuteGas()
-	msgJson["sender"] = msg.GetSender()
-}
-
-func decodeMsgReportData(msg *types.MsgReportData, msgJson common.JsDict) {
-	msgJson["request_id"] = msg.GetRequestID()
-	msgJson["raw_reports"] = msg.GetRawReports()
-	msgJson["validator"] = msg.GetValidator()
-	msgJson["reporter"] = msg.GetReporter()
-}
-
-func decodeMsgCreateDataSource(msg *types.MsgCreateDataSource, msgJson common.JsDict) {
-	msgJson["name"] = msg.GetName()
-	msgJson["description"] = msg.GetDescription()
-	msgJson["executable"] = msg.GetExecutable()
-	msgJson["fee"] = msg.GetFee()
-	msgJson["treasury"] = msg.GetTreasury()
-	msgJson["owner"] = msg.GetOwner()
-	msgJson["sender"] = msg.GetSender()
-}
-
-func decodeMsgCreateOracleScript(msg *types.MsgCreateOracleScript, msgJson common.JsDict) {
-	msgJson["name"] = msg.GetName()
-	msgJson["description"] = msg.GetDescription()
-	msgJson["schema"] = msg.GetSchema()
-	msgJson["source_code_url"] = msg.GetSourceCodeURL()
-	msgJson["code"] = msg.GetCode()
-	msgJson["owner"] = msg.GetOwner()
-	msgJson["sender"] = msg.GetSender()
-}
-
-func decodeMsgEditDataSource(msg *types.MsgEditDataSource, msgJson common.JsDict) {
-	msgJson["data_source_id"] = msg.GetDataSourceID()
-	msgJson["name"] = msg.GetName()
-	msgJson["description"] = msg.GetDescription()
-	msgJson["executable"] = msg.GetExecutable()
-	msgJson["fee"] = msg.GetFee()
-	msgJson["treasury"] = msg.GetTreasury()
-	msgJson["owner"] = msg.GetOwner()
-	msgJson["sender"] = msg.GetSender()
-}
-
-func decodeMsgEditOracleScript(msg *types.MsgEditOracleScript, msgJson common.JsDict) {
-	msgJson["oracle_script_id"] = msg.GetOracleScriptID()
-	msgJson["name"] = msg.GetName()
-	msgJson["description"] = msg.GetDescription()
-	msgJson["schema"] = msg.GetSchema()
-	msgJson["source_code_url"] = msg.GetSourceCodeURL()
-	msgJson["code"] = msg.GetCode()
-	msgJson["owner"] = msg.GetOwner()
-	msgJson["sender"] = msg.GetSender()
-}
-
-func decodeMsgAddReporter(msg *types.MsgAddReporter, msgJson common.JsDict) {
-	msgJson["validator"] = msg.GetValidator()
-	msgJson["reporter"] = msg.GetReporter()
-}
-
-func decodeMsgRemoveReporter(msg *types.MsgRemoveReporter, msgJson common.JsDict) {
-	msgJson["validator"] = msg.GetValidator()
-	msgJson["reporter"] = msg.GetReporter()
-}
-
-func decodeMsgActivate(msg *types.MsgActivate, msgJson common.JsDict) {
-	msgJson["validator"] = msg.GetValidator()
-}
